requirements.txt
sseclient-py
confluent-kafka
urllib3
streamlit
plotly
pinotdb
pandas
requests
dash
watchdog
pulsar-client